src/requirements.txt
flask==2.3.3
python-jose[cryptography]==3.3.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv